asist bot/requirements.txt
aiogram==2.25.1
pytz==2023.3
SQLAlchemy==2.0.23
python-dotenv==1.0.0